IEBH-438: added first pilot-hdc service (#1)
In this commit we are adding the first pilot-hdc service: `metadata`.
It requires `kafka` and `postgres`, so we are adding those services too.

The `postgres` init script should be placed in
`helm_charts/postgres/postgres-init.sql`. It is git ignored due to
security concerns.

A previous version of `metadata` required vaul too, so we added it.
But we then pushed a new version of the `metadata` service that does NOT
require vault and made the vault deployment optional.

We might remove it if it's not needed once we have all services running.